+++ src/server.R
@@ -8,19 +8,32 @@ server <- function(input, output) {
     person <- person(
       sex = as.numeric(input$sex), height = input$height, bmi = input$bmi
     )
+    if (input$group == "simple") {
+      # Workaround while plots are not implemented for all groups
+      vo2_ml_min_results <- group$vo2_ml_min(group, person)
+    } else {
+      vo2_ml_min_results <- matrix(c(group$vo2_ml_min(group, person), NA, NA), 1)
+    }
     data.frame(
       "Metric" = c(
         "VO2 ml/min", "VO2 ml/kg/min", "Heart rate", "Ventilation",
         "Oxygen pulse", "VE/VCO2 slope", "Breathing frequency"
       ),
       "Value" = c(
-        group$vo2_ml_min(group, person)[1],
+        vo2_ml_min_results[1, 1],
         group$vo2_ml_kg_min(group, person),
         group$heart_rate(group, person),
         group$ventilation(group, person),
         group$oxygen_pulse(group, person),
         group$ve_vco2_slope(group, person),
         group$breathing_frequency(group, person)
+      ),
+      "Confidence_Interval" = c(
+        paste0(
+          round(vo2_ml_min_results[1, 2], 2), " - ",
+          round(vo2_ml_min_results[1, 3], 2)
+        ),
+        NA, NA, NA, NA, NA, NA
       )
     )
   })