+++ src/methods.R
@@ -37,6 +37,7 @@ expand_matrix <- function(mx, times) {
 }
 
 y_hat_ci <- function(x, metric, metric_data, transf) {
+  x <- matrix(x, ncol = 2L) # Workaround for when x is c()
   beta_hat <- metric_data$beta_hat[[metric]]
   sigma_beta_hat <- metric_data$sigma_beta_hat[[metric]]
   weights <- metric_data$haukeland_vyntus
@@ -55,10 +56,13 @@ y_hat_ci <- function(x, metric, metric_data, transf) {
   t(mat)
 }
 
-create_x <- function(person, x_function) {
+create_x <- function(person, sex, x_expr) {
+  expr <- substitute(x_expr)
+  # Build a function of person and sex with expr as the body
+  x_function <- eval(bquote(function(person, sex) .(expr)))
   vapply(
     X = c(person$sex, 1 - person$sex),
-    FUN = function (sex) x_function(person, sex),
+    FUN = function(sex) x_function(person, sex),
     FUN.VALUE = numeric(length(x_function(person, sex)))
   )
 }