+++ NEWS.md
@@ -1,3 +1,7 @@
+# (under development)
+
+* Calculation of confidence intervals (issue #11)
+
 # Deployment v0.0.0.9003
 
 - Fixed calculation when only "haukeland" or "vyntus" is used (issue #12)